# build/app/swagger.json
{
    "swagger": "2.0",
    "host": "35.233.176.30:8081",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/hello": {
            "get": {
                "description": "Returns Hello World",
                "responses": {
                    "200": {
                        "description": "Hello World",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the state of the SYS node running in VPS and some additional system data.",
                "responses": {
                    "200": {
                        "description": "Returns general VPS status",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notarize": {
            "post": {
                "description": "Returns the state of the SYS node running in VPS and some additional system data.",
                "responses": {
                    "200": {
                        "description": "Returns general VPS status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Returns error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Returns error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Returns error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "body",
                        "name": "tx",
                        "required": true,
                        "type": "string",
                        "description": "transaction hex to be notarized"
                    }
                ]
            }
        },
        "/blacklist": {
            "get": {
                "description": "Returns the blacklist",
                "responses": {
                    "200": {
                        "description": "Returns current blacklisted addresses",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds an address entry to the blacklist",
                "responses": {
                    "200": {
                        "description": "Returns status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "parameters": [
                    {
                        "in": "body",
                        "name": "address",
                        "required": true,
                        "type": "string",
                        "description": "address to be added to the blacklist",
                        "schema": {
                            "type": "object",
                            "required": [
                                "address"
                            ],
                            "properties": {
                                "address": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ]
            },
            "delete": {
                "description": "Deletes an address entry in the blacklist",
                "responses": {
                    "200": {
                        "description": "Returns status",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "body",
                        "name": "address",
                        "required": true,
                        "type": "string",
                        "description": "address to be removed from the blacklist",
                        "schema": {
                            "type": "object",
                            "required": [
                                "address"
                            ],
                            "properties": {
                                "address": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ]
            }
        },
        "/notarization-errors": {
            "get": {
                "description": "returns a JSON array of the notarization error objects in the DB",
                "responses": {
                    "200": {
                        "description": "Returns general VPS status",
                        "schema": {
                            "type": "array"
                        }
                    }
                }
            }
        }
    }
}